started WorkingScheduleBusinessCalendarUiTest #40
