+++ src/test/groovy/io/jmix/tests/ui/screen/administration/businesscalendars/editor/BusinessCalendarEditor.groovy
@@ -3,6 +3,7 @@ package io.jmix.tests.ui.screen.administration.businesscalendars.editor
 import io.jmix.masquerade.Wire
 import io.jmix.masquerade.base.Composite
 import io.jmix.masquerade.component.Button
+import io.jmix.masquerade.component.TabSheet
 import io.jmix.masquerade.component.TextField
 
 class BusinessCalendarEditor extends Composite<BusinessCalendarEditor> {
@@ -15,6 +16,9 @@ class BusinessCalendarEditor extends Composite<BusinessCalendarEditor> {
 
     @Wire
     TextField sourceField
+
+    @Wire
+    TabSheet workingScheduleTabId
 
     @Wire
     Button create
